requirements.txt
#cmake>=3.30
dlib==19.24.6
face_recognition>=1.3
fastapi[standard]>=0.115.2
google-cloud-storage>=2.18.2